Cast

+++ Firestore/core/test/unit/model/values_test.cc
@@ -120,12 +120,12 @@ class ValuesTest : public ::testing::Test {
         EXPECT_EQ(expected_result, Values::Compare(val1, val2))
             << "Order check failed for '" << Values::CanonicalId(val1)
             << "' and ' " << Values::CanonicalId(val2) << " (expected "
-            << expected_result << ")";
+            << static_cast<int>(expected_result) << ")";
         EXPECT_EQ(util::ReverseOrder(expected_result),
                   Values::Compare(val2, val1))
             << "Reverse order check failed for '" << Values::CanonicalId(val1)
             << "' and ' " << Values::CanonicalId(val2) << " (expected "
-            << util::ReverseOrder(expected_result) << ")";
+            << static_cast<int>(util::ReverseOrder(expected_result)) << ")";
       }
     }
   }
